+++ spring-boot-with-jsp/boiler-plate-proj/src/main/java/com/example/boilerplateproj/domain/board/entity/Board.java
@@ -17,7 +17,7 @@ public class Board {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.AUTO)
-	private int boardNo;
+	private long boardNo;
 	private String title;
 	private String content;
 	private String writer;
